集成FastDFS

+++ xmall-sellergoods-service/src/main/java/cn/jerio/sellergoods/service/impl/TypeTemplateServiceImpl.java
@@ -125,7 +125,7 @@ public class TypeTemplateServiceImpl implements TypeTemplateService {
 			//查询规格选项列表
 			TbSpecificationOptionExample example=new TbSpecificationOptionExample();
 			TbSpecificationOptionExample.Criteria criteria = example.createCriteria();
-			criteria.andSpecIdEqualTo((Long) map.get("id"));
+			criteria.andSpecIdEqualTo(Long.valueOf((Integer)map.get("id")));
 			List<TbSpecificationOption> options = specificationOptionMapper.selectByExample(example);
 			map.put("options", options);
 		}
